missions/week 1/back/requirements.txt
Django==3.1.3
mysqlclient==2.0.1
django-mysql==3.9
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
requests==2.22.0
django-cors-headers==3.2.1
elasticsearch